Rename two_phase module to two_phase_commit
