+++ services/scabbard/libscabbard/src/store/scabbard_store/diesel/operations/add_consensus_event.rs
@@ -35,7 +35,7 @@ use crate::store::scabbard_store::diesel::{
 use crate::store::scabbard_store::ScabbardStoreError;
 use crate::store::scabbard_store::{
     event::ConsensusEvent,
-    two_phase::{Event, Message},
+    two_phase_commit::{Event, Message},
 };
 
 use super::ScabbardStoreOperations;
