+++ services/scabbard/libscabbard/src/store/scabbard_store/diesel/operations/list_consensus_events.rs
@@ -33,7 +33,7 @@ use crate::store::scabbard_store::diesel::{
 use crate::store::scabbard_store::ScabbardStoreError;
 use crate::store::scabbard_store::{
     event::IdentifiedConsensusEvent,
-    two_phase::{Event, Message},
+    two_phase_commit::{Event, Message},
 };
 
 use super::ScabbardStoreOperations;
