+++ services/scabbard/libscabbard/src/store/scabbard_store/event.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::store::scabbard_store::two_phase::Event;
+use crate::store::scabbard_store::two_phase_commit::Event;
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum ConsensusEvent {
